x86/split_lock: Move Split and Bus lock code to a dedicated file
mainline inclusion
from v6.11-rc2
commit 350afa8a1101f62ce31bc4ed6f69cf4b90ec4fa2
category: feature

-------------------

Bus Lock Detect functionality on AMD platforms works identical to Intel.

Move split_lock and bus_lock specific code from intel.c to a dedicated
file so that it can be compiled and supported on non-Intel platforms.

Also, introduce CONFIG_X86_BUS_LOCK_DETECT, make it dependent on
CONFIG_CPU_SUP_INTEL and add compilation dependency of the new bus_lock.c
file on CONFIG_X86_BUS_LOCK_DETECT.

Link: https://lore.kernel.org/all/20240808062937.1149-2-ravi.bangoria@amd.com
